refactor: 컴포넌트 분리
상품 주문 내역 페이지의 코드가 한 파일에 몰려 있어서 가독성이 좋지 않은 것 같아 분리
